# openapi/openapi.generated.yaml
openapi: 3.0.0
info:
  title: Generated routes
  version: 1.0.0
paths:
...
  "/api/posts/{postId}/comments":
    get:
          summary: Generated stub (GET)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/profile/photo":
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/profile/posts":
    delete:
          summary: Generated stub (DELETE)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
  "/api/tree/view-preference":
    patch:
          summary: Generated stub (PATCH)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
  "/api/trust/check-opportunity":
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/trust/respond":
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
  "/api/users/lookup-by-email":
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
  "/api/waitlist":
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error